Fourth week Lab and Assignment Problems

+++ src/week2/PraticeProblems/AsciiOps.java
@@ -22,7 +22,7 @@ public class AsciiOps {
 
     static int[] toAsciiArray(String s) {
         int[] arr = new int[s.length()];
-        for (int i=0;i<s.length();i++) arr[i] = (int)s.charAt(i);
+        for (int i=0;i<s.length();i++) arr[i] = s.charAt(i);
         return arr;
     }
